TD2 graph database design/app/requirements.txt
fastapi
uvicorn
psycopg2-binary
neo4j

